♻️ Refactor : organisms
