--- src/feature/picsel/shared/components/ui/organisms/PicselBookSelect.tsx
@@ -1,13 +1,0 @@
-import React from 'react';
-
-import { View, Text } from 'react-native';
-
-const PicselBookSelect = () => {
-  return (
-    <View>
-      <Text>PicselBookSelect</Text>
-    </View>
-  );
-};
-
-export default PicselBookSelect;
