+++ src/screens/picsel/picselBook/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import PicselBookTabContent from '@/feature/picsel/shared/components/ui/organisms/PicselBookTabContent';
-import PicselBookTabHeader from '@/feature/picsel/shared/components/ui/organisms/PicselBookTabHeader';
+import PicselBookTabContent from '@/feature/picsel/shared/components/layouts/PicselBookTabContent';
+import PicselBookTabHeader from '@/feature/picsel/shared/components/layouts/PicselBookTabHeader';
 import { usePicselBookTab } from '@/feature/picsel/shared/hooks/usePicselBookTab';
 import ScreenLayout from '@/shared/components/layouts/ScreenLayout';
